🔨[refactor] 서비스 로직 단순 주석 추가 / common - response 패키지 명 error로 수정
