+++ src/main/java/com/daon/backend/common/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.daon.backend.common.response;
+package com.daon.backend.common.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
