+++ src/main/java/com/daon/backend/config/SecurityConfig.java
@@ -1,5 +1,9 @@
-package com.daon.backend.security;
+package com.daon.backend.config;
 
+import com.daon.backend.security.JwtAuthenticationEntryPoint;
+import com.daon.backend.security.JwtAuthenticationFilter;
+import com.daon.backend.security.JwtAuthenticationProvider;
+import com.daon.backend.security.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
