+++ src/main/java/com/daon/backend/image/service/EmptyImageException.java
@@ -7,8 +7,4 @@ public class EmptyImageException extends AbstractException {
     public EmptyImageException() {
         super("빈 이미지 파일입니다.");
     }
-
-    public EmptyImageException(String message) {
-        super(message);
-    }
 }
