+++ src/main/java/com/daon/backend/image/service/ImageFileService.java
@@ -1,5 +1,6 @@
 package com.daon.backend.image.service;
 
+import com.daon.backend.image.dto.UploadedImage;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface ImageFileService {
