+++ src/main/java/com/daon/backend/member/controller/MemberErrorHandler.java
@@ -1,8 +1,8 @@
 package com.daon.backend.member.controller;
 
 import com.daon.backend.common.exception.DomainSpecificAdvice;
-import com.daon.backend.common.response.ErrorCode;
-import com.daon.backend.common.response.ErrorResponse;
+import com.daon.backend.common.error.ErrorCode;
+import com.daon.backend.common.error.ErrorResponse;
 import com.daon.backend.member.domain.EmailNotFoundException;
 import com.daon.backend.member.domain.MemberNotFoundException;
 import com.daon.backend.member.domain.PasswordMismatchException;
