+++ src/main/java/com/daon/backend/member/service/MemberService.java
@@ -22,6 +22,9 @@ public class MemberService {
     private final SessionMemberProvider sessionMemberProvider;
     private final MemberServiceThroughTask memberServiceThroughTask;
 
+    /**
+     * 회원 가입
+     */
     @Transactional
     public void signUp(SignUpRequestDto requestDto) {
         memberRepository.findByUsername(requestDto.getUsername())
@@ -40,6 +43,9 @@ public class MemberService {
         );
     }
 
+    /**
+     * 회원 정보 수정
+     */
     @Transactional
     public void modifyMember(ModifyMemberRequestDto requestDto) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -53,6 +59,9 @@ public class MemberService {
                 );
     }
 
+    /**
+     * 회원(본인) 정보 조회
+     */
     public FindMemberResponseDto findMember() {
         String memberId = sessionMemberProvider.getMemberId();
         Member member = memberRepository.findById(memberId)
@@ -61,12 +70,18 @@ public class MemberService {
         return new FindMemberResponseDto(member);
     }
 
+    /**
+     * 회원 검색
+     */
     public SearchMemberResponseDto searchMember(String username) {
         List<MemberSummary> memberSummaries = memberRepository.searchMembersByUsername(username);
 
         return new SearchMemberResponseDto(memberSummaries);
     }
 
+    /**
+     * 이메일 추가
+     */
     @Transactional
     public void createEmail(AddEmailRequestDto requestDto) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -77,6 +92,9 @@ public class MemberService {
         member.createEmail(email);
     }
 
+    /**
+     * 이메일 목록 조회
+     */
     public FindEmailsResponseDto findEmails() {
         String memberId = sessionMemberProvider.getMemberId();
         Member findmember = memberRepository.findById(memberId)
@@ -89,6 +107,9 @@ public class MemberService {
         );
     }
 
+    /**
+     * 이메일 삭제
+     */
     @Transactional
     public void deleteEmail(Long memberEmailId) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -96,7 +117,10 @@ public class MemberService {
                 .orElseThrow(() -> MemberNotFoundException.byMemberId(memberId));
         member.removeEmail(memberEmailId);
     }
-  
+
+    /**
+     * 회원 탈퇴
+     */
     public void withdrawMember() {
         String memberId = sessionMemberProvider.getMemberId();
         Member member = memberRepository.findById(memberId)
