+++ src/main/java/com/daon/backend/task/domain/workspace/CanNotDeletePersonalWorkspaceException.java
@@ -2,9 +2,9 @@ package com.daon.backend.task.domain.workspace;
 
 import com.daon.backend.common.exception.AbstractException;
 
-public class CanNotPersonalWorkspaceException extends AbstractException {
+public class CanNotDeletePersonalWorkspaceException extends AbstractException {
 
-    public CanNotPersonalWorkspaceException(Long workspaceId) {
+    public CanNotDeletePersonalWorkspaceException(Long workspaceId) {
         super("개인 워크스페이스는 삭제할 수 없습니다. workspaceId: " + workspaceId);
     }
 }
