+++ src/main/java/com/daon/backend/task/service/TaskService.java
@@ -19,6 +19,9 @@ public class TaskService {
     private final TaskRepository taskRepository;
     private final SessionMemberProvider sessionMemberProvider;
 
+    /**
+     * 할 일 생성
+     */
     @Transactional
     public CreateTaskResponseDto createTask(Long projectId, CreateTaskRequestDto requestDto) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -52,6 +55,9 @@ public class TaskService {
         return new CreateTaskResponseDto(taskId);
     }
 
+    /**
+     * 할 일 단 건 조회
+     */
     public FindTaskResponseDto findTask(Long taskId) {
         return new FindTaskResponseDto(
                 taskRepository.findTaskDetail(sessionMemberProvider.getMemberId(), taskId)
@@ -59,6 +65,9 @@ public class TaskService {
         );
     }
 
+    /**
+     * 할 일 수정
+     */
     @Transactional
     public void modifyTask(Long projectId, Long taskId, ModifyTaskRequestDto requestDto) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -88,6 +97,9 @@ public class TaskService {
         );
     }
 
+    /**
+     * 북마크 설정/해제
+     */
     @Transactional
     public SetBookmarkResponseDto setBookmark(Long projectId, Long taskId) {
         String memberId = sessionMemberProvider.getMemberId();
@@ -113,6 +125,9 @@ public class TaskService {
         return new SetBookmarkResponseDto(created);
     }
 
+    /**
+     * 할 일 진행 상태 변경
+     */
     @Transactional
     public void modifyTaskProgressStatus(Long projectId, Long taskId, ModifyProgressStatusRequestDto requestDto) {
         Task task = taskRepository.findTaskByTaskId(taskId)
